LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/items/row-minimum-size-grid-items-003-expected.txt


FAIL .masonry 1 assert_equals:
<div class="masonry" style="grid-template-rows: auto auto;">
  <div class="verticalLR" data-expected-height="100">XXXXXXXXXX</div>
  <div data-expected-height="100"></div>
</div>
height expected 100 but got 0
PASS .masonry 2
FAIL .masonry 3 assert_equals:
<div class="masonry" style="grid-template-rows: 20px 20px;">
  <div class="verticalLR" data-expected-height="40">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
PASS .masonry 4
FAIL .masonry 5 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="40">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 6 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 7 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 8 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="0" style="margin: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 9 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="30" style="margin: 5px 0px;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 10 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="padding: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 11 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="40" style="padding: 5px 0px;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 12 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="border: solid 5px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 13 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="40" style="border: solid 5px blue;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 14 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 15 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 16 assert_equals:
<div class="masonry" style="grid-template-rows: auto auto;">
  <div class="verticalRL" data-expected-height="100">XXXXXXXXXX</div>
  <div data-expected-height="100"></div>
</div>
height expected 100 but got 0
PASS .masonry 17
FAIL .masonry 18 assert_equals:
<div class="masonry" style="grid-template-rows: 20px 20px;">
  <div class="verticalRL" data-expected-height="40">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
PASS .masonry 19
FAIL .masonry 20 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="40">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 21 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 22 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 23 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="0" style="margin: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 24 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="30" style="margin: 5px 0px;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 25 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="padding: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 26 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="40" style="padding: 5px 0px;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 27 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="border: solid 5px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 28 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="40" style="border: solid 5px blue;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 29 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .masonry 30 assert_equals:
<div class="masonry" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="40"></div>
</div>
height expected 40 but got 0
FAIL .masonry 31 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: auto auto;">
  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>
  <div data-expected-width="100"></div>
</div>
width expected 100 but got 0
PASS .masonry 32
FAIL .masonry 33 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: 20px 20px;">
  <div class="horizontalTB" data-expected-width="40">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
PASS .masonry 34
FAIL .masonry 35 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 36 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 37 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 38 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 39 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="30" style="margin: 0px 5px;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 40 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 41 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40" style="padding: 0px 5px;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 42 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 43 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40" style="border: solid 5px blue;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 44 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 45 assert_equals:
<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 46 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: auto auto;">
  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>
  <div data-expected-width="100"></div>
</div>
width expected 100 but got 0
PASS .masonry 47
FAIL .masonry 48 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: 20px 20px;">
  <div class="horizontalTB" data-expected-width="40">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
PASS .masonry 49
FAIL .masonry 50 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 51 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 52 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 53 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 54 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="30" style="margin: 0px 5px;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 55 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 56 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40" style="padding: 0px 5px;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 57 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 58 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="40" style="border: solid 5px blue;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
FAIL .masonry 59 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .masonry 60 assert_equals:
<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="40"></div>
</div>
width expected 40 but got 0
item writing-mode: vertical-lr;

grid-template-rows: auto auto;
XXXXXXXXXX
grid-template-rows: 0px 0px;
XXXXXXXXXX
grid-template-rows: 20px 20px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item margin height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item margin height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item padding height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item padding height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item border height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item border height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item height + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item height + margin + border + padding: 10px;
item writing-mode: vertical-rl;

grid-template-rows: auto auto;
XXXXXXXXXX
grid-template-rows: 0px 0px;
XXXXXXXXXX
grid-template-rows: 20px 20px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item margin height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item margin height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item padding height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item padding height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item border height: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item border height: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item height + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item height + margin + border + padding: 10px;
masonry container writing-mode: vertical-lr; & item writing-mode: horizontal-tb;

grid-template-rows: auto auto;
XXXXXXXXXX
grid-template-rows: 0px 0px;
XXXXXXXXXX
grid-template-rows: 20px 20px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item margin width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item margin width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item padding width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item padding width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item border width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item border width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item width + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item width + margin + border + padding: 10px;
masonry container writing-mode: vertical-rl; & item writing-mode: horizontal-tb;

grid-template-rows: auto auto;
XXXXXXXXXX
grid-template-rows: 0px 0px;
XXXXXXXXXX
grid-template-rows: 20px 20px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item margin width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item margin width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item padding width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item padding width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item border width: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item border width: 10px;
grid-template-rows: minmax(auto, 0px) minmax(auto, 0px); item width + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 20px) minmax(auto, 20px); item width + margin + border + padding: 10px;
